[ContactsUI] Unify ContactsUI.Mac into ContactsUI.
It's the only framework that uses a different api definition (and makefile
name), so merge these to make upcoming code that automatically generates the
list of frameworks per platform (based on the makefiles) simpler.
