+++ src/frameworks.sources
@@ -1520,7 +1520,7 @@ MAC_FRAMEWORKS =            \
 	CFNetwork               \
 	CloudKit                \
 	Contacts                \
-	ContactsUI.Mac		\
+	ContactsUI              \
 	CoreAnimation           \
 	CoreAudioKit            \
 	CoreBluetooth           \
